pyproject.toml
[tool.poetry]
name = "caillou"
version = "0.0.1"
description = "AI-powered toolbox for the terminal"
authors = ["Sylvain <sdudoit@proton.me>"]
license = "GPLv3"
readme = "README.md"

homepage = "https://github.com/sdudoit/caillou"
repository = "https://github.com/sdudoit/caillou"

[tool.poetry.dependencies]
python = "^3.11"
click = "^8.1.7"
langchain = "^0.1.16"
python-dotenv = "^1.0.1"
openai = "^1.21.2"
langchain-openai = "^0.1.3"
textual = "^0.57.1"
clipboard = "^0.0.4"
platformdirs = "^4.2.1"

[tool.poetry.group.dev.dependencies]
black = "^24.4.0"
isort = "^5.13.2"
mypy = "^1.9.0"
ruff = "^0.4.0"
textual-dev = "^1.5.1"

[tool.black]
line-length = 120
exclude = '''
/(
    \.git
  | \.mypy_cache
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
force_single_line = true
atomic = true
lines_after_imports = 2
lines_between_types = 1

[tool.ruff]
select = ["E", "F", "W", "Q", "I"]
ignore = ["E203"]
fixable = ["ALL"]
unfixable = []
line-length = 120


[tool.mypy]
disallow_untyped_defs = true
show_error_codes = true
no_implicit_optional = true
warn_return_any = true
warn_unused_ignores = true
exclude = ["docs", "tests", "examples", "tools", "build"]


[tool.poetry.scripts]
caillou = 'caillou.__main__:main'
caillou_translator = 'caillou.apps.translator:main'

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

